api/b_idt.c: B_IdtExecuteIdts() => "unstatify" because of pb in C++ API api/iodebase.h: B_IdtExecuteIdts() declaration restored api/k_emain.c, k_graph.c, k_ibst.c; k_idif.c, k_wks.c: various cast to avoid warnings + code cleanup

+++ api/k_emain.c
@@ -151,7 +151,7 @@ char *EXP_addprepost(char* pre, char* post, char* src, char** tg)
 
     buf = SCR_stracpy(src);
     SCR_strip(buf);
-    *tg = SW_nalloc(strlen(buf) +
+    *tg = SW_nalloc((int)strlen(buf) +
                     (pre == NULL ? 0 : (int)strlen(pre)) +
                     (post == NULL ? 0 : (int)strlen(post)) + 1);
     if(*tg != NULL)
